# docs/swagger.yaml
basePath: /
definitions:
...
  models.CommentStruct:
    properties:
      comments:
        items:
          properties:
            author:
              properties:
                email:
                  type: string
                id:
                  type: string
                name:
                  type: string
              type: object
            createdTime:
              type: string
            id:
              type: string
            lastUpdatedTime: {}
            mentioned:
              properties:
                usrL2PNC5o3H4lBEi:
                  properties:
                    displayName:
                      type: string
                    email:
                      type: string
                    id:
                      type: string
                    type:
                      type: string
                  type: object
              type: object
            text:
              type: string
          type: object
        type: array
      offset: {}
    type: object
...
host: localhost:8012
info:
  contact:
    email: eti@cisco.io
    name: API Support
  description: This is a Kosha REST serice for exposing many airtable features as
    REST APIs with better consistency, observability etc
  termsOfService: http://swagger.io/terms/
  title: Airtable Connector API
  version: "2.0"
paths:
...
  /api/v0/{baseId}/{tableIdOrName}/{recordId}/comments:
    get:
      consumes:
      - application/json
      description: Get Comments
      parameters:
      - description: Base Id
        in: path
        name: baseId
        required: true
        type: string
      - description: Table Id
        in: path
        name: tableIdOrName
        required: true
        type: string
      - description: Base Id
        in: path
        name: recordId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CommentStruct'
      summary: Get airtable
      tags:
      - Comments
...
  /api/v0/specification/list:
    get:
      consumes:
      - application/json
      description: Get all environment variables that need to be supplied
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: bad request
          schema:
            type: string
        "403":
          description: permission denied
          schema:
            type: string
        "404":
          description: not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get details of the connector specification
      tags:
      - specification
swagger: "2.0"